Create class4 file in data package
* Remove class4 functions from utils.misc
