+++ utils/misc.py
@@ -1,22 +1,22 @@
-import fcntl
+import datetime
 import os
+import re
 import time
-from flask import current_app
 import xml.etree.ElementTree as ET
-from shapely.geometry.polygon import LinearRing
-from netCDF4 import Dataset, chartostring
+from operator import itemgetter
+
 import cftime
 import numpy as np
+import pyproj
+from flask import current_app
+from netCDF4 import Dataset, chartostring
 from shapely.geometry import LineString, Point, Polygon
 from shapely.geometry.multipolygon import MultiPolygon
-from thredds_crawler.crawl import Crawl
-import datetime
-import pyproj
-from operator import itemgetter
-from oceannavigator import DatasetConfig
-import re
+from shapely.geometry.polygon import LinearRing
+
 from data import open_dataset
-import pickle as pickle
+from oceannavigator import DatasetConfig
+
 
 def list_kml_files(subdir):
     DIR = os.path.join(current_app.config['OVERLAY_KML_DIR'], subdir)
@@ -82,7 +82,7 @@ def points(file_id, projection, resolution, extent):
             lonlat = list(map(float, c_txt.split(',')))
             x, y = proj(lonlat[0], lonlat[1])
             p = Point(y, x)
-            
+
             if view.envelope.intersects(p):
                 points.append({
                     'type': "Feature",
@@ -167,18 +167,18 @@ def list_areas(file_id, simplify=True):
                         continue
                     lonlat = tup.split(',')
                     coords.append([float(lonlat[1]), float(lonlat[0])])
-                
+
                 if simplify:
                     coords = list(LinearRing(coords).simplify(1.0 / 32).coords)
-                
+
                 result.append(coords)
-            
+
         return result
 
     for place in folder.iter(nsmap + "Placemark"):
         outers = get_coords("outerBoundaryIs")
         inners = get_coords("innerBoundaryIs")
-        
+
         name = None
         for placename in place.iter(nsmap + "name"):
             name = placename.text
@@ -190,7 +190,7 @@ def list_areas(file_id, simplify=True):
             'centroids': [(c.y, c.x) for c in centroids],
             'key': file_id + "/" + name,
         })
-    
+
     areas = sorted(areas, key=lambda k: k['name'])
 
     return areas
@@ -225,9 +225,9 @@ def areas(area_id, projection, resolution, extent):
                     ).astype(float)
                     ix, iy = proj(lonlat_inner[:, 0], lonlat_inner[:, 1])
                     holes.append(list(zip(iy, ix)))
-    
+
             polys.append(Polygon(list(zip(oy, ox)), holes))
-        
+
         mp = MultiPolygon(polys).simplify(resolution * 1.5)
 
         def get_coordinates(poly):
@@ -239,7 +239,7 @@ def areas(area_id, projection, resolution, extent):
                 inn = np.array(i.coords)
                 coords.append(
                     np.array(proj(inn[:, 0], inn[:, 1],
-                             inverse=True)).transpose().tolist())
+                                  inverse=True)).transpose().tolist())
             return coords
 
         if view.envelope.intersects(mp):
@@ -511,217 +511,6 @@ def drifters_time(drifter_id):
         'min': min_time.isoformat(),
         'max': max_time.isoformat(),
     }
-
-
-def list_class4_files_slowly():
-    # This function has poor performance; only use as a fallback.
-    c = Crawl(current_app.config["CLASS4_CATALOG_URL"], select=[".*_GIOPS_.*.nc$"],
-              workers=16)
-
-    result = []
-    for dataset in c.datasets:
-        value = dataset.name[:-3]
-        date = datetime.datetime.strptime(value.split("_")[1], "%Y%m%d")
-        result.append({
-            'name': date.strftime("%Y-%m-%d"),
-            'id': value
-        })
-
-    return result
-
-
-def list_class4_files():
-    """Return the list of Class 4 files.
-
-    For performance reasons, this function reads the list of Class 4 files
-    from a cache file containing the predetermined list. If the cache file is
-    not accessible, this function falls back to the slow method of generating
-    the list on the fly.
-    """
-    cache_file_name = os.path.join(current_app.config['CACHE_DIR'],
-                                   'class4_files.pickle')
-    try:
-        fp = open(cache_file_name, 'rb')
-    except IOError as e:
-        msg = ('Warning: Unable to open cache list of Class 4 files: %s\n'
-               'Falling back to slow method for generating the list of '
-               'Class 4 files on the fly')
-        msg = msg % (str(e),)
-        print(msg)
-        return list_class4_files_slowly()
-
-    # We need to read from the cache file. To ensure another process is not
-    # currently *writing* to the cache file, first acquire a shared lock (i.e.,
-    # a read lock) on the file. We make at most "max_tries" attempts to acquire
-    # the lock.
-    max_tries = 10
-    num_tries = 0
-    attempt_lock = True
-    while attempt_lock:
-        try:
-            fcntl.lockf(fp, fcntl.LOCK_SH | fcntl.LOCK_NB)
-        except IOError:
-            num_tries += 1
-            if num_tries == max_tries:
-                lock_acquired = False
-                attempt_lock = False
-            else:
-                time.sleep(1)
-        else:
-            lock_acquired = True
-            attempt_lock = False
-
-    if lock_acquired:
-        result = pickle.load(fp)
-        fcntl.lockf(fp, fcntl.LOCK_UN)
-    else:
-        msg = ('Warning: Unable to acquire read lock on cache file\n'
-               'Falling back to slow method for generating list of Class 4 '
-               'files on the fly')
-        print(msg)
-        result = list_class4_files_slowly()
-    fp.close()
-
-    return result
-
-
-def list_class4(d):
-    dataset_url = current_app.config["CLASS4_URL"] % d
-
-    with Dataset(dataset_url, 'r') as ds:
-        lat = ds['latitude'][:]
-        lon = ds['longitude'][:]
-        ids = list(map(str.strip, chartostring(ds['id'][:])))
-        rmse = []
-
-        for i in range(0, lat.shape[0]):
-            best = ds['best_estimate'][i, 0, :]
-            obsv = ds['observation'][i, 0, :]
-            rmse.append(np.ma.sqrt(((best - obsv) ** 2).mean()))
-
-    rmse = np.ma.hstack(rmse)
-    maxval = rmse.mean() + 2 * rmse.std()
-    rmse_norm = rmse / maxval
-
-    loc = list(zip(lat, lon))
-
-    points = []
-    for idx, ll in enumerate(loc):
-        if np.ma.is_masked(rmse[idx]):
-            continue
-        points.append({
-            'name': "%s" % ids[idx],
-            'loc': "%f,%f" % ll,
-            'id': "%s/%d" % (d, idx),
-            'rmse': float(rmse[idx]),
-            'rmse_norm': float(rmse_norm[idx]),
-        })
-        points = sorted(points, key=lambda k: k['id'])
-
-    return points
-
-
-def class4(class4_id, projection, resolution, extent):
-    dataset_url = current_app.config["CLASS4_URL"] % class4_id
-
-    proj = pyproj.Proj(init=projection)
-    view = _get_view(extent)
-
-    rmse = []
-    lat = []
-    lon = []
-    point_id = []
-    identifiers = []
-    with Dataset(dataset_url, 'r') as ds:
-        lat_in = ds['latitude'][:]
-        lon_in = ds['longitude'][:]
-        ids = list(map(str.strip, chartostring(ds['id'][:])))
-
-        for i in range(0, lat_in.shape[0]):
-            x, y = proj(lon_in[i], lat_in[i])
-            p = Point(y, x)
-            if view.envelope.intersects(p):
-                lat.append(float(lat_in[i]))
-                lon.append(float(lon_in[i]))
-                identifiers.append(ids[i])
-                best = ds['best_estimate'][i, 0, :]
-                obsv = ds['observation'][i, 0, :]
-                point_id.append(i)
-                rmse.append(np.ma.sqrt(((best - obsv) ** 2).mean()))
-
-    rmse = np.ma.hstack(rmse)
-    rmse_norm = np.clip(rmse / 1.5, 0, 1)
-
-    loc = list(zip(lon, lat))
-
-    points = []
-
-    for idx, ll in enumerate(loc):
-        if np.ma.is_masked(rmse[idx]):
-            continue
-        points.append({
-            'type': "Feature",
-            'geometry': {
-                'type': "Point",
-                'coordinates': ll,
-            },
-            'properties': {
-                'name': "%s" % identifiers[idx],
-                'id': "%s/%d" % (class4_id, point_id[idx]),
-                'error': float(rmse[idx]),
-                'error_norm': float(rmse_norm[idx]),
-                'type': 'class4',
-                'resolution': 0,
-            },
-        })
-
-    result = {
-        'type': "FeatureCollection",
-        'features': points,
-    }
-
-    return result
-
-
-def list_class4_forecasts(class4_id):
-    dataset_url = current_app.config["CLASS4_URL"] % class4_id
-    with Dataset(dataset_url, 'r') as ds:
-        var = ds['modeljuld']
-        forecast_date = [d.strftime("%d %B %Y") for d in
-                         cftime.utime(var.units).num2date(var[:])]
-
-    res = [{
-        'id': 'best',
-        'name': 'Best Estimate',
-    }]
-
-    if len(set(forecast_date)) > 1:
-        for idx, date in enumerate(forecast_date):
-            if res[-1]['name'] == date:
-                continue
-            res.append({
-                'id': idx,
-                'name': date
-            })
-
-    return res
-
-
-def list_class4_models(class4_id):
-    select = ["(.*/)?%s.*_profile.nc$" % class4_id[:16]]
-    c = Crawl(current_app.config["CLASS4_CATALOG_URL"], select=select)
-
-    result = []
-    for dataset in c.datasets:
-        value = dataset.name[:-3]
-        model = value.split("_")[2]
-        if model != "GIOPS":
-            result.append({
-                'value': value.split("_")[2],
-                'id': value
-            })
-
-    return result
 
 
 def get_point_data(dataset, variable, time, depth, location):
